build: move lambdas to a nested stack
Ref. metriport/metriport-internal#1040
