+++ packages/infra/lib/lambdas-nested-stack.ts
@@ -1,0 +1,348 @@
+import { Duration, NestedStack, NestedStackProps } from "aws-cdk-lib";
+import { SnsAction } from "aws-cdk-lib/aws-cloudwatch-actions";
+import * as ec2 from "aws-cdk-lib/aws-ec2";
+import * as lambda from "aws-cdk-lib/aws-lambda";
+import { Function as Lambda } from "aws-cdk-lib/aws-lambda";
+import * as rds from "aws-cdk-lib/aws-rds";
+import * as s3 from "aws-cdk-lib/aws-s3";
+import * as secret from "aws-cdk-lib/aws-secretsmanager";
+import { Construct } from "constructs";
+import { EnvConfig } from "../config/env-config";
+import { EnvType } from "./env-type";
+import { createLambda } from "./shared/lambda";
+import { LambdaLayers, setupLambdasLayers } from "./shared/lambda-layers";
+import { Secrets } from "./shared/secrets";
+
+export const CDA_TO_VIS_TIMEOUT = Duration.minutes(15);
+
+interface LambdasNestedStackProps extends NestedStackProps {
+  config: EnvConfig;
+  vpc: ec2.IVpc;
+  secrets: Secrets;
+  dbCluster: rds.IDatabaseCluster;
+  dbCredsSecret: secret.ISecret;
+  medicalDocumentsBucket: s3.Bucket;
+  sandboxSeedDataBucket: s3.IBucket | undefined;
+  alarmAction?: SnsAction;
+}
+
+export class LambdasNestedStack extends NestedStack {
+  readonly lambdaLayers: LambdaLayers;
+  readonly cdaToVisualizationLambda: Lambda;
+  readonly documentDownloaderLambda: lambda.Function;
+  readonly outboundPatientDiscoveryLambda: lambda.Function;
+  readonly outboundDocumentQueryLambda: lambda.Function;
+  readonly outboundDocumentRetrievalLambda: lambda.Function;
+
+  constructor(scope: Construct, id: string, props: LambdasNestedStackProps) {
+    super(scope, id, props);
+
+    this.lambdaLayers = setupLambdasLayers(this);
+
+    this.cdaToVisualizationLambda = this.setupCdaToVisualization({
+      lambdaLayers: this.lambdaLayers,
+      vpc: props.vpc,
+      envType: props.config.environmentType,
+      medicalDocumentsBucket: props.medicalDocumentsBucket,
+      sandboxSeedDataBucket: props.sandboxSeedDataBucket,
+      sentryDsn: props.config.lambdasSentryDSN,
+      alarmAction: props.alarmAction,
+    });
+
+    this.documentDownloaderLambda = this.setupDocumentDownloader({
+      lambdaLayers: this.lambdaLayers,
+      vpc: props.vpc,
+      secrets: props.secrets,
+      cwOrgCertificate: props.config.cwSecretNames.CW_ORG_CERTIFICATE,
+      cwOrgPrivateKey: props.config.cwSecretNames.CW_ORG_PRIVATE_KEY,
+      bucketName: props.medicalDocumentsBucket.bucketName,
+      envType: props.config.environmentType,
+      sentryDsn: props.config.lambdasSentryDSN,
+    });
+
+    this.outboundPatientDiscoveryLambda = this.setupOutboundPatientDiscovery({
+      lambdaLayers: this.lambdaLayers,
+      vpc: props.vpc,
+      envType: props.config.environmentType,
+      sentryDsn: props.config.lambdasSentryDSN,
+      alarmAction: props.alarmAction,
+      dbCluster: props.dbCluster,
+      dbCredsSecret: props.dbCredsSecret,
+      // TODO move this to a config
+      maxPollingDuration: Duration.minutes(11),
+    });
+
+    this.outboundDocumentQueryLambda = this.setupOutboundDocumentQuery({
+      lambdaLayers: this.lambdaLayers,
+      vpc: props.vpc,
+      envType: props.config.environmentType,
+      sentryDsn: props.config.lambdasSentryDSN,
+      alarmAction: props.alarmAction,
+      dbCluster: props.dbCluster,
+      dbCredsSecret: props.dbCredsSecret,
+      // TODO move this to a config
+      maxPollingDuration: Duration.minutes(15),
+    });
+
+    this.outboundDocumentRetrievalLambda = this.setupOutboundDocumentRetrieval({
+      lambdaLayers: this.lambdaLayers,
+      vpc: props.vpc,
+      envType: props.config.environmentType,
+      sentryDsn: props.config.lambdasSentryDSN,
+      alarmAction: props.alarmAction,
+      dbCluster: props.dbCluster,
+      dbCredsSecret: props.dbCredsSecret,
+      // TODO move this to a config
+      maxPollingDuration: Duration.minutes(15),
+    });
+  }
+
+  private setupCdaToVisualization(ownProps: {
+    lambdaLayers: LambdaLayers;
+    vpc: ec2.IVpc;
+    envType: EnvType;
+    medicalDocumentsBucket: s3.Bucket;
+    sandboxSeedDataBucket: s3.IBucket | undefined;
+    sentryDsn: string | undefined;
+    alarmAction: SnsAction | undefined;
+  }): Lambda {
+    const {
+      lambdaLayers,
+      vpc,
+      sentryDsn,
+      envType,
+      alarmAction,
+      medicalDocumentsBucket,
+      sandboxSeedDataBucket,
+    } = ownProps;
+
+    const cdaToVisualizationLambda = createLambda({
+      stack: this,
+      name: "CdaToVisualization",
+      runtime: lambda.Runtime.NODEJS_16_X,
+      entry: "cda-to-visualization",
+      envType,
+      envVars: {
+        CDA_TO_VIS_TIMEOUT_MS: CDA_TO_VIS_TIMEOUT.toMilliseconds().toString(),
+        ...(sentryDsn ? { SENTRY_DSN: sentryDsn } : {}),
+      },
+      layers: [lambdaLayers.shared, lambdaLayers.chromium],
+      memory: 1024,
+      timeout: CDA_TO_VIS_TIMEOUT,
+      vpc,
+      alarmSnsAction: alarmAction,
+    });
+
+    medicalDocumentsBucket.grantReadWrite(cdaToVisualizationLambda);
+
+    if (sandboxSeedDataBucket) {
+      sandboxSeedDataBucket.grantReadWrite(cdaToVisualizationLambda);
+    }
+
+    return cdaToVisualizationLambda;
+  }
+
+  /**
+   * We are intentionally not setting an alarm action for this lambda, as many issues
+   * may be caused outside of our system. To eliminate noise, we will not alarm on this
+   * lambda.
+   */
+  private setupDocumentDownloader(ownProps: {
+    lambdaLayers: LambdaLayers;
+    vpc: ec2.IVpc;
+    secrets: Secrets;
+    cwOrgCertificate: string;
+    cwOrgPrivateKey: string;
+    bucketName: string | undefined;
+    envType: EnvType;
+    sentryDsn: string | undefined;
+  }): Lambda {
+    const {
+      lambdaLayers,
+      vpc,
+      secrets,
+      cwOrgCertificate,
+      cwOrgPrivateKey,
+      bucketName,
+      envType,
+      sentryDsn,
+    } = ownProps;
+
+    const documentDownloaderLambda = createLambda({
+      stack: this,
+      name: "DocumentDownloader",
+      runtime: lambda.Runtime.NODEJS_18_X,
+      entry: "document-downloader",
+      envType,
+      envVars: {
+        TEST_ENV: "TEST",
+        CW_ORG_CERTIFICATE: cwOrgCertificate,
+        CW_ORG_PRIVATE_KEY: cwOrgPrivateKey,
+        ...(bucketName && {
+          MEDICAL_DOCUMENTS_BUCKET_NAME: bucketName,
+        }),
+        ...(sentryDsn ? { SENTRY_DSN: sentryDsn } : {}),
+      },
+      layers: [lambdaLayers.shared],
+      memory: 512,
+      timeout: Duration.minutes(5),
+      vpc,
+    });
+
+    // granting secrets read access to lambda
+    const cwOrgCertificateKey = "CW_ORG_CERTIFICATE";
+    if (!secrets[cwOrgCertificateKey]) {
+      throw new Error(`${cwOrgCertificateKey} is not defined in config`);
+    }
+    secrets[cwOrgCertificateKey].grantRead(documentDownloaderLambda);
+
+    const cwOrgPrivateKeyKey = "CW_ORG_PRIVATE_KEY";
+    if (!secrets[cwOrgPrivateKeyKey]) {
+      throw new Error(`${cwOrgPrivateKeyKey} is not defined in config`);
+    }
+    secrets[cwOrgPrivateKeyKey].grantRead(documentDownloaderLambda);
+
+    return documentDownloaderLambda;
+  }
+
+  private setupOutboundPatientDiscovery(ownProps: {
+    lambdaLayers: LambdaLayers;
+    vpc: ec2.IVpc;
+    envType: EnvType;
+    dbCredsSecret: secret.ISecret;
+    dbCluster: rds.IDatabaseCluster;
+    maxPollingDuration: Duration;
+    sentryDsn: string | undefined;
+    alarmAction: SnsAction | undefined;
+  }): Lambda {
+    const {
+      lambdaLayers,
+      dbCredsSecret,
+      vpc,
+      sentryDsn,
+      envType,
+      alarmAction,
+      dbCluster,
+      maxPollingDuration,
+    } = ownProps;
+
+    const outboundPatientDiscoveryLambda = createLambda({
+      stack: this,
+      name: "OutboundPatientDiscovery",
+      entry: "ihe-outbound-patient-discovery",
+      envType,
+      envVars: {
+        ...(sentryDsn ? { SENTRY_DSN: sentryDsn } : {}),
+        DB_CREDS: dbCredsSecret.secretArn,
+        MAX_POLLING_DURATION: maxPollingDuration
+          .minus(Duration.minutes(1))
+          .toMilliseconds()
+          .toString(),
+      },
+      layers: [lambdaLayers.shared],
+      memory: 512,
+      timeout: maxPollingDuration,
+      vpc,
+      alarmSnsAction: alarmAction,
+    });
+
+    dbCluster.connections.allowDefaultPortFrom(outboundPatientDiscoveryLambda);
+    dbCredsSecret.grantRead(outboundPatientDiscoveryLambda);
+
+    return outboundPatientDiscoveryLambda;
+  }
+
+  private setupOutboundDocumentQuery(ownProps: {
+    lambdaLayers: LambdaLayers;
+    vpc: ec2.IVpc;
+    envType: EnvType;
+    dbCredsSecret: secret.ISecret;
+    dbCluster: rds.IDatabaseCluster;
+    maxPollingDuration: Duration;
+    sentryDsn: string | undefined;
+    alarmAction: SnsAction | undefined;
+  }): Lambda {
+    const {
+      lambdaLayers,
+      dbCredsSecret,
+      vpc,
+      sentryDsn,
+      envType,
+      alarmAction,
+      dbCluster,
+      maxPollingDuration,
+    } = ownProps;
+
+    const outboundDocumentQueryLambda = createLambda({
+      stack: this,
+      name: "OutboundDocumentQuery",
+      entry: "ihe-outbound-document-query",
+      envType,
+      envVars: {
+        ...(sentryDsn ? { SENTRY_DSN: sentryDsn } : {}),
+        DB_CREDS: dbCredsSecret.secretArn,
+        MAX_POLLING_DURATION: maxPollingDuration
+          .minus(Duration.minutes(1))
+          .toMilliseconds()
+          .toString(),
+      },
+      layers: [lambdaLayers.shared],
+      memory: 512,
+      timeout: maxPollingDuration,
+      vpc,
+      alarmSnsAction: alarmAction,
+    });
+
+    dbCluster.connections.allowDefaultPortFrom(outboundDocumentQueryLambda);
+    dbCredsSecret.grantRead(outboundDocumentQueryLambda);
+
+    return outboundDocumentQueryLambda;
+  }
+
+  private setupOutboundDocumentRetrieval(ownProps: {
+    lambdaLayers: LambdaLayers;
+    vpc: ec2.IVpc;
+    envType: EnvType;
+    dbCredsSecret: secret.ISecret;
+    dbCluster: rds.IDatabaseCluster;
+    maxPollingDuration: Duration;
+    sentryDsn: string | undefined;
+    alarmAction: SnsAction | undefined;
+  }): Lambda {
+    const {
+      lambdaLayers,
+      dbCredsSecret,
+      vpc,
+      sentryDsn,
+      envType,
+      alarmAction,
+      dbCluster,
+      maxPollingDuration,
+    } = ownProps;
+
+    const outboundDocumentRetrievalLambda = createLambda({
+      stack: this,
+      name: "OutboundDocumentRetrieval",
+      entry: "ihe-outbound-document-retrieval",
+      envType,
+      envVars: {
+        ...(sentryDsn ? { SENTRY_DSN: sentryDsn } : {}),
+        DB_CREDS: dbCredsSecret.secretArn,
+        MAX_POLLING_DURATION: maxPollingDuration
+          .minus(Duration.minutes(1))
+          .toMilliseconds()
+          .toString(),
+      },
+      layers: [lambdaLayers.shared],
+      memory: 512,
+      timeout: maxPollingDuration,
+      vpc,
+      alarmSnsAction: alarmAction,
+    });
+
+    dbCluster.connections.allowDefaultPortFrom(outboundDocumentRetrievalLambda);
+    dbCredsSecret.grantRead(outboundDocumentRetrievalLambda);
+
+    return outboundDocumentRetrievalLambda;
+  }
+}
